added preliminary support for lidar data input.  'pipeline 2' now includes an option to reduce the elevation data resolution before processing.  bug fixes to java tools to support processing of larger elevation data sets.

+++ catchment-delineation-helpers/src/main/java/ca/bc/gov/catchment/algorithms/NearestNeighbour.java
@@ -34,7 +34,7 @@ public class NearestNeighbour {
 		this(features, 200);
 	}
 	
-	public NearestNeighbour(SimpleFeatureSource features, int initialSearchRadius) {
+	public NearestNeighbour(SimpleFeatureSource features, double initialSearchRadius) {
 		this.features = features;
 		this.initialSearchRadiusMetres = initialSearchRadius;
